newTrack

+++ app/routes/track.py
@@ -62,14 +62,17 @@ def save_layer(trackID):
 @login_required
 def del_layer(trackID,layerID):
 	track = Track.objects().get(id = ObjectId(trackID))	
-	track.update(pull__layers__layerID=ObjectId(layerID))
-	track.save()
+	if current_user.username == track.createdBy:
+		track.update(pull__layers__layerID=ObjectId(layerID))
+		track.save()
 
-	if len(track.layers)-1 == 0: #why is -1 necessary?
-		print "Deleting Track"
-		track.delete()
-		return redirect( url_for('track.all_tracks') )
+		if len(track.layers)-1 == 0: #why is -1 necessary?
+			track.delete()
+			return redirect( url_for('track.all_tracks') )
 	return redirect( url_for('track.track_page', trackID=trackID))
+
+
+# @track.route('delete/<trackID>')
 
 
 @track.route('/new', methods = ["POST","GET"])
